Patch from Manuel Ruiz to make admin caommands easily callable from within code.
Refactor regarding creation of Context (new
InitialContext, getInitialContext in Ra....) into a Singleton class
(InitialContextBuilder) and if no environment is given or it is null, will try to
locate a ejbca.properties to extract from it the jndi properties
before using the default one.
Also make the ouputstream settable, with default beeing System.out.


git-svn-id: https://svn.cesecore.eu/svn/ejbca/trunk@2043 37aeb1de-b14b-0410-9a0e-82b28170d228